+++ src/main/java/name/neuhalfen/projects/crypto/bouncycastle/openpgp/algorithms/PGPCompressionAlgorithms.java
@@ -2,8 +2,6 @@ package name.neuhalfen.projects.crypto.bouncycastle.openpgp.algorithms;
 
 import org.bouncycastle.bcpg.CompressionAlgorithmTags;
 
-
-@SuppressWarnings("PMD.ShortVariable")
 public enum PGPCompressionAlgorithms {
   /**
    * No compression.
@@ -22,9 +20,13 @@ public enum PGPCompressionAlgorithms {
    */
   BZIP2(CompressionAlgorithmTags.BZIP2);
 
-  public final int id;
+  private final int algorithmId;
 
-  PGPCompressionAlgorithms(int id) {
-    this.id = id;
+  public int getAlgorithmId() {
+    return algorithmId;
+  }
+
+  PGPCompressionAlgorithms(int algorithmId) {
+    this.algorithmId = algorithmId;
   }
 }